Made a property: public Mark currentplayer that sets the currentPlayer to PlayerX (initial player turn). Used within method PlaceMark.
Wrote method ChangePlayerTurn that checks who the currentPlayer is and changes into the other player depending on current value. This method is also called within method "PlaceMark" since we need to shift player turn when a player's turn is over.

Wrote method IsBoardFull, that calls method IsFree to check all the fields on the board to see if there are any free fields. If no empty fields are found, method returns true, else false.