Changed some stuff in the controller methods (controller class GameSessionController), added two separate action methods- one that takes care of showing the gameboard: "ShowGameBoard" and one that calls the method "PlaceMark" in the Game Model: "PlaceMark". To be able to show an updated version of the Gameboard after a player has placed a mark on it, we return a call to the first action method in the ActionResult PlaceMark.
Also made changes in View-class ShowGameBoard to be able to use the Action methods in the controller to push data to the view:
- Created helper method "ShowMarkAt" that calls a method in the model (Game class) called "GetMarkAt". This method gets the Mark that has been placed on a specific field on the gameboard and prints out the corresponding text value depending on whether it was player X or player O that placed his mark on the field.

In the Html that follows, where we declare the buttons we then call the helper method to show the correct player mark on each button (field).

Method mentioned above that was added to the Game class:
GetMarkAt, this method just returns the Mark (PlayerX or PlayerO) that was found on the specific x and y coordinates.