Show winner in view and prevent new moves if there is a winner.
Modified method PlaceMark in Game class so it checks both that the HasWinner-method is not true and that the IsFree method is true for a player to be able to place a mark on the board. This because when there is a winner we don't want the game to go on.
The method HasWinner is called within the view class ShowGameBoard, because we wan't to print out who is the winner according to the result of this method call.