Enabled and created joining of games.
Properties added in Gamesession class:
GameFull- a bool that returns true if number of players in a specific gamesession is 2.

State - enum that defins two states for a gamesession, a game can either be in "waiting" state or "Started". Waiting state means that 1 player has created a game, the game can't be started until 2 players are present.
Started state means that 2 players has joined the session and the gamesession is therefore in "started" state.

PlayersInSpecificGame - A list where the player-objects for a specific gamesession will be stored.

Methods added in GameSession class:
GameSession constructor- used to instansiate a new gamesession for two users.

JoinGame (player p)-
A user can join a gamesession if the gamesession is not yet full (only has 1 player so far) and if the currentstate of the gamesession is "Waiting". Then the player 'p' is added to the list "PlayersInSpecificGame".

StartGame-
If the game has 2 players (GameFull ==true) and the currentstate of the game is Waiting, then the currentstate is changed to Started.

GameOver- Method that returns true or false depending on if a gamesession is over or not. A GameSession is over if there are not 2 players in the gamesessiuon and if the currentstate of the gamesession is Started or if the specific Game for the gamesession has a winner.

Methods added in GameSessionController:
CreateGameForm-
Method that handles the request when a user wants to create a new game session, if a user (a specific browser) already has an ongoing gamesession the request is re-directed to the gamesession that is already running. Else- we loop through the dictionary over all gamesessions to check if there is a game where there is just 1 player so far. If a gamesession is found with only 1 player in it, the game is added to the list of "open games", meaning: games that are listed on the start page that users can join.

JoinGame- Here we instansiate a second player object, the player will be playerO since here "join game" means joining an existing game. We also say that the id we get through the method call should be looked up in the dictionary to find the gamesession with that id, this is assigned to a game reference. This game reference is then used to call the method JoinGame in the GameSession class, also the method StartGame in GameSession class is called. The Session key also is assigned the value secondPlayer.

CreateGame- Action method that takes care of the request and response for the creation of a new gamesession. A new Gamesession ID is generated, and a new GameSession is instansiated. The key (gameID) and the value (newGame) is added to the Dictionary GameSessions. Also her we add the first player object to the game session.

CreateGameForm
-Fixed so that the start page has two options:
either the user can choose to create a new game, or the user can choose to join a game that is in the list of open games. There are a nickname textfield by each of these options.